+++ backend/src/map/map.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Param, Query } from '@nestjs/common';
+import { Controller, Get, Query } from '@nestjs/common';
 import { MapService } from './map.service';
 
 @Controller('places')